+++ src/commands/builtins/builtin_help.c
@@ -6,7 +6,7 @@
 #include "config/config.h"
 
 static void print_builtin(const Builtin *builtin) {
-    shell_print(SHELL_INFO,
+    shell_print(SHELL_HELP,
         "\nCommand: %s\n"
         "-----------------\n"
         "Usage:\n  %s\n\n"
@@ -14,19 +14,19 @@ static void print_builtin(const Builtin *builtin) {
         builtin->name, builtin->usage, builtin->description);
 
     if (builtin->flags && *builtin->flags) {
-        shell_print(SHELL_INFO, "Flags:\n");
+        shell_print(SHELL_HELP, "Flags:\n");
         for (const char **f = builtin->flags; *f != NULL; ++f) {
-            shell_print(SHELL_INFO, "  %s\n", *f);
+            shell_print(SHELL_HELP, "  %s\n", *f);
         }
-        shell_print(SHELL_INFO, "\n");
+        shell_print(SHELL_HELP, "\n");
     }
 
     if (builtin->examples && *builtin->examples) {
-        shell_print(SHELL_INFO, "Examples:\n");
+        shell_print(SHELL_HELP, "Examples:\n");
         for (const char **e = builtin->examples; *e != NULL; ++e) {
-            shell_print(SHELL_INFO, "  %s\n", *e);
+            shell_print(SHELL_HELP, "  %s\n", *e);
         }
-        shell_print(SHELL_INFO, "\n");
+        shell_print(SHELL_HELP, "\n");
     }
 }
 
@@ -46,12 +46,12 @@ static void print_envs() {
         if (len > max_key_len) max_key_len = len;
     }
 
-    shell_print(SHELL_INFO, "\nEnvironment Variables:\n-----------------\n");
+    shell_print(SHELL_HELP, "\nEnvironment Variables:\n-----------------\n");
 
     if (env_fields_count > 0) {
-        shell_print(SHELL_INFO, "\nDefault Environment Variables:\n-----------------\n");
+        shell_print(SHELL_HELP, "\nDefault Environment Variables:\n-----------------\n");
         for (size_t i = 0; i < env_fields_count; ++i) {
-            shell_print(SHELL_INFO, "%-*s = %-20s  (%s)\n",
+            shell_print(SHELL_HELP, "%-*s = %-20s  (%s)\n",
                         (int)max_key_len,
                         env_fields[i].key,
                         env_fields[i].value,
@@ -59,7 +59,7 @@ static void print_envs() {
         }
     }
 
-    shell_print(SHELL_INFO, "\nCustom Environment Variables:\n-----------------\n");
+    shell_print(SHELL_HELP, "\nCustom Environment Variables:\n-----------------\n");
     for (size_t i = 0; i < shell_config.env_count; ++i) {
         bool is_default = false;
         for (size_t j = 0; j < env_fields_count; ++j) {
@@ -70,7 +70,7 @@ static void print_envs() {
         }
         if (is_default) continue;
 
-        shell_print(SHELL_INFO, "%-*s = %s\n",
+        shell_print(SHELL_HELP, "%-*s = %s\n",
                     (int)max_key_len,
                     shell_config.envs[i].key,
                     shell_config.envs[i].value);
@@ -90,9 +90,9 @@ static void print_alias() {
     }
 
     if (alias_fields_count > 0) {
-        shell_print(SHELL_INFO, "\nDefault Aliases:\n-----------------\n");
+        shell_print(SHELL_HELP, "\nDefault Aliases:\n-----------------\n");
         for (size_t i = 0; i < alias_fields_count; ++i) {
-            shell_print(SHELL_INFO, "%-*s = %-20s  (%s)\n",
+            shell_print(SHELL_HELP, "%-*s = %-20s  (%s)\n",
                         (int)max_key_len,
                         alias_fields[i].key,
                         alias_fields[i].value,
@@ -100,7 +100,7 @@ static void print_alias() {
         }
     }
 
-    shell_print(SHELL_INFO, "\nCustom Aliases:\n-----------------\n");
+    shell_print(SHELL_HELP, "\nCustom Aliases:\n-----------------\n");
     for (size_t i = 0; i < shell_config.alias_count; ++i) {
         bool is_default = false;
         for (size_t j = 0; j < alias_fields_count; ++j) {
@@ -111,11 +111,87 @@ static void print_alias() {
         }
         if (is_default) continue;
 
-        shell_print(SHELL_INFO, "%-*s = %s\n",
+        shell_print(SHELL_HELP, "%-*s = %s\n",
                     (int)max_key_len,
                     shell_config.aliases[i].key,
                     shell_config.aliases[i].value);
     }
+}
+
+size_t find_matching_configs(const char *arg, const ConfigField **out_matches, size_t max_matches) {
+    size_t count = 0;
+    for (size_t i = 0; i < num_config && count < max_matches; ++i) {
+        const ConfigField *field = &config_fields[i];
+        if (strcmp(field->name, arg) == 0) {
+            out_matches[count++] = field;
+            continue;
+        }
+        if (field->aliases) {
+            for (const char **alias = field->aliases; *alias != NULL; ++alias) {
+                if (strcmp(*alias, arg) == 0) {
+                    out_matches[count++] = field;
+                    break;
+                }
+            }
+        }
+    }
+    return count;
+}
+
+static void print_config(const char *arg, const ConfigField *field, size_t index) {
+    if (index == 0) {
+        shell_print(SHELL_HELP, "\nMatching Configs for '%s':\n-----------------\n", arg);
+    }
+    shell_print(SHELL_HELP,
+        "\nConfig %zu: %s\n"
+        "-----------------\n"
+        "Description:\n  %s\n\n",
+        index + 1, field->name, field->description);
+
+    const char *default_value = get_default_config_field(field->name);
+    const char *current_value = get_config_field(field->name);
+
+    if (current_value) shell_print(SHELL_HELP, "Current value: %s\n", current_value);
+    if (default_value) shell_print(SHELL_HELP, "Default value: %s\n", default_value);
+    if (current_value || default_value) shell_print(SHELL_HELP, "\n");
+
+    if (field->valid_options && *field->valid_options) {
+        shell_print(SHELL_HELP, "Valid options:\n");
+        for (const char **option = field->valid_options; *option != NULL; ++option) {
+            shell_print(SHELL_HELP, "  %s\n", *option);
+        }
+        shell_print(SHELL_HELP, "\n");
+    }
+
+    if (field->aliases && *field->aliases) {
+        shell_print(SHELL_HELP, "Help Aliases:\n");
+        for (const char **alias = field->aliases; *alias != NULL; ++alias) {
+            shell_print(SHELL_HELP, "  %s\n", *alias);
+        }
+        shell_print(SHELL_HELP, "\n");
+    }
+
+    shell_print(SHELL_HELP, "Usage:\n  set %s <value>\n\n", field->name);
+}
+
+// TODO: make good looking
+void print_all_config() {
+    shell_print(SHELL_HELP, "\nAvailable Config Fields:\n-----------------------\n");
+    for (size_t i = 0; i < num_config; ++i) {
+        const ConfigField *field = &config_fields[i];
+        shell_print(SHELL_HELP, "%s", field->name);
+
+        if (field->aliases && *field->aliases) {
+            shell_print(SHELL_HELP, " (aliases: ");
+            for (const char **alias = field->aliases; *alias != NULL; ++alias) {
+                shell_print(SHELL_HELP, "%s", *alias);
+                if (*(alias + 1) != NULL) shell_print(SHELL_HELP, ", ");
+            }
+            shell_print(SHELL_HELP, ")");
+        }
+        shell_print(SHELL_HELP, "\n");
+    }
+    shell_print(SHELL_HELP, "\n");
 }
 
 void builtin_help(int argc, const char **argv) {
@@ -138,6 +214,20 @@ void builtin_help(int argc, const char **argv) {
 
     if (strcmp(arg, "alias") == 0) {
         print_alias();
+        return;
+    }
+
+    if (strcmp(arg, "config") == 0) {
+        print_all_config();
+        return;
+    }
+
+    const ConfigField *matches[32];
+    size_t config_matches = find_matching_configs(arg, matches, 32);
+    if (config_matches > 0) {
+        for (size_t i = 0; i < config_matches; ++i) {
+            print_config(arg, matches[i], i);
+        }
         return;
     }